# perf(system): store value instead of version number in Link for dirty check