@medz 2024-12-21
@johnsoncodehk Will the performance be reduced when value encounters complex objects? Dynamic languages ​​do not use RuntimeType and memory address judgment like statically compiled languages.