@medz 2024-12-21
Additionally, the range of boost seems to be between normal dynamic fluctuations.