@medz 2024-12-21
https://github.com/stackblitz/alien-signals/blob/deprecated-version/src/system.ts#L138
https://github.com/stackblitz/alien-signals/blob/deprecated-version/src/system.ts#L172

update to `>> SubscriberFlags.CanPropagate`

ref #13 

Easier to understand