@medz 2024-12-21
👍
<img width="696" alt="image" src="https://github.com/user-attachments/assets/a5c9cf41-fecb-42be-8652-0af1f3295703" />
I just wrote a bench for complex objects, and it really improves things.