@johnsoncodehk 2024-12-21
> I just wrote a bench for complex objects, and it really improves things.

Thanks for the information, could you open a PR for add this bench?