@johnsoncodehk 2024-12-21
> update to `>> SubscriberFlags.CanPropagate`

That's coincidental, but actually incorrect, consider if `SubscriberFlags` were as follows:

```ts
export const enum SubscriberFlags {
	None = 0,
	Tracking = 1 << 10,
	CanPropagate = 1 << 11,
	RunInnerEffects = 1 << 12,
	ToCheckDirty = 1 << 13,
	Dirty = 1 << 14,
}
```

The code should be `>> 12` instead of `>> SubscriberFlags.CanPropagate`. (`SubscriberFlags.CanPropagate` is 4096)